moved style to separate fil
