+++ packages/boba/gateway/src/containers/Bridging/BridgeInput/EmergencySwap/index.tsx
@@ -1,4 +1,3 @@
-import { Button, Typography } from 'components/global'
 import React, { FC, useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import {
@@ -8,47 +7,15 @@ import {
   selectlayer2Balance,
 } from 'selectors'
 import networkService from 'services/networkService'
-import styled from 'styled-components'
+
+import { getETHMetaTransaction } from 'actions/setupAction'
+import { openAlert } from 'actions/uiAction'
+import BN from 'bignumber.js'
+import { isEqual } from 'lodash'
+import { logAmount } from 'util/amountConvert'
 import { LAYER } from 'util/constant'
 import { NETWORK } from 'util/network/network.util'
-import BN from 'bignumber.js'
-import { logAmount } from 'util/amountConvert'
-import { isEqual } from 'lodash'
-import { openAlert } from 'actions/uiAction'
-import { getETHMetaTransaction } from 'actions/setupAction'
-
-export const SwapContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  padding: 10px;
-  gap: 8px;
-  border-radius: 8px;
-  border: 1px solid
-    ${({ theme: { colors, name } }) =>
-      name === 'light' ? colors.gray[600] : colors.green[300]};
-  background: ${({ theme: { name } }) =>
-    name === 'light' ? 'transparant' : 'rgba(238, 238, 238, 0.05)'};
-`
-export const SwapAlert = styled(Typography).attrs({
-  variant: 'body3',
-})`
-  font-weight: 400;
-  line-height: normal;
-  color: ${({ theme, color }) =>
-    color
-      ? color
-      : theme.name === 'light'
-      ? theme.colors.gray[700]
-      : theme.colors.gray[100]};
-`
-export const SwapAction = styled(Button)`
-  padding: 10px;
-  display: flex;
-  flex-direction: column;
-  align-item: center;
-  justify-content: center;
-`
+import { SwapAction, SwapAlert, SwapContainer } from './styles'
 
 interface Props {}
